Moved `data_thread` from autonomous to data module
